feat: write JSON to a file
In some cases, if a list of stale repositories becomes sufficiently long, dealing with it as an output instead of a file is cumbersome. When using the data in the GHA output, you may start to see errors such as "Argument list too long" because the output needs to be passed to the next step's invocation. This change adds a "stale_repos.json" file that will contain the stale repos JSON data for use in further processing.