fix: go back to root user in Dockefile
Fixes: #101

Based on [GitHub docs]() we will not be able to access the workspace
of the GitHub Action without being the root user.  As a non-root user
we won't be able to write to `$GITHUB_OUTPUT` which is an environment
variable that is a path inside the workspace and GitHub Actions
using to handle output from the GitHub Action.

Once that was realized, this seems to be the only possible path.

fix: ignore checkov linter requiring user in Dockerfile

Signed-off-by: jmeridth <jmeridth@gmail.com>